Created Database connection
Got Database connection working (Yay!). In a very rudimentary form. (Not so Yay)

Data base connection is set up in NodeJS folder. This folder is set up as its own application running server.js as the main file. Not all current folders are necessarily relevant. I will be doing additional clean up this weekend.

Important Note: Node(database) app has to be started separately from the React app.(UI) in order for it to work. This can be done by opening a terminal from the NodeJS files and running the following command:
node server.js

The only working query currently runs from the Tracking webpage, in the Yellow box. Ouput of the query can be seen in the terminal windows of the NodeJS terminal.